requirements.txt
APScheduler==3.11.0
discord.py==2.3.2
PyNaCl==1.5.0
imageio_ffmpeg==0.6.0
python-dotenv==1.1.1
validators==0.18.2
yt_dlp==2025.7.21
redis>=5.0.0
async-timeout>=4.0.3,<5
aiohttp==3.9.5
pytz